[#328] Move data security provision into a component
