+++ frontend/src/components/utils/index.js
@@ -18,3 +18,4 @@ export const checkPasswordCriteria = (value) => {
 export { default as SaveAsImageButton } from "./SaveAsImageButton";
 export { default as PasswordCriteria } from "./PasswordCriteria";
 export { default as ShowLabelButton } from "./ShowLabelButton";
+export { default as DataSecurityProvisionModal } from "./DataSecurityProvisionModal";
